OSIS-1826 refactored tests
