+++ program_management/tests/ddd/service/read/test_get_end_postponement_year_service.py
@@ -23,113 +23,19 @@
 # ############################################################################
 from unittest import mock
 
-from django.test import SimpleTestCase, TestCase
+from django.test import SimpleTestCase
 
-from base.tests.factories.academic_year import AcademicYearFactory
-from education_group.tests.ddd.factories.repository.fake import get_fake_training_repository, \
-    get_fake_mini_training_repository
-from education_group.tests.ddd.factories.training import TrainingFactory
-from education_group.tests.factories.mini_training import MiniTrainingFactory
-from program_management.ddd import command
-from program_management.ddd.domain.program_tree_version import STANDARD
 from program_management.ddd.domain.service.calculate_end_postponement import CalculateEndPostponement
-from program_management.ddd.service.read import node_identity_service
+from program_management.ddd.service.read import get_end_postponement_year_service
 from program_management.tests.ddd.factories.commands.get_end_postponement_year_command import \
     GetEndPostponementYearCommandFactory
-from program_management.tests.ddd.factories.program_tree import ProgramTreeFactory
-from program_management.tests.ddd.factories.program_tree_version import ProgramTreeVersionFactory
-from program_management.tests.ddd.factories.repository.fake import get_fake_program_tree_repository, \
-    get_fake_program_tree_version_repository
-from testing.mocks import MockPatcherMixin
 
 
-class TestNodeIdentityService(MockPatcherMixin, TestCase):
-    def setUp(self):
-        self.current_year = 2020
-        self.maximum_postponement_year = 2025
-        AcademicYearFactory.produce_in_future(self.current_year)
+class TestGetEndPostponementYearService(SimpleTestCase):
 
-        self.root_node_code = 'LDROI200M'
-
-        self.program_tree = ProgramTreeFactory(
-            entity_id__year=self.current_year,
-            entity_id__code=self.root_node_code,
-            # root_node__end_year=self.current_year,
+    @mock.patch.object(CalculateEndPostponement, 'calculate_program_tree_end_postponement')
+    def test_domain_service_is_called(self, mock_domain_service):
+        get_end_postponement_year_service.calculate_program_tree_end_postponement(
+            GetEndPostponementYearCommandFactory()
         )
-        self.fake_program_tree_repository = get_fake_program_tree_repository([self.program_tree])
-
-        self.program_tree_version = ProgramTreeVersionFactory(
-            tree=self.program_tree,
-            entity_id__version_name=STANDARD,
-            program_tree_repository=self.fake_program_tree_repository
-        )
-        self.fake_program_tree_version_repository = get_fake_program_tree_version_repository([self.program_tree_version])
-
-        self.command = GetEndPostponementYearCommandFactory(
-            code=self.root_node_code,
-            year=self.current_year,
-        )
-
-    # @mock.patch("program_management.ddd.domain.service.identity_search._get_data_from_db")
-    # def test_when_root_node_is_minitraining_and_end_year_lt_maximum_postponement(self, mock_data_from_db):
-    #     end_year = self.maximum_postponement_year - 3
-    #     mini_training = MiniTrainingFactory(entity_identity__year=self.current_year, end_year=self.current_year)
-    #     mock_data_from_db.return_value = {"offer_acronym": mini_training.entity_identity.acronym, "offer_type": mini_training.type.name}
-    #     fake_minitraining_repository = get_fake_mini_training_repository([mini_training])
-    #     result = CalculateEndPostponement.calculate_program_tree_end_postponement(
-    #         identity=self.program_tree.entity_id,
-    #         training_repository=None,
-    #         mini_training_repository=fake_minitraining_repository
-    #     )
-    #     self.assertEqual(result, end_year)
-
-    @mock.patch("program_management.ddd.domain.service.identity_search._get_data_from_db")
-    def test_when_root_node_is_training_and_end_year_is_none(self, mock_data_from_db):
-        training = TrainingFactory(entity_identity__year=self.current_year, end_year=None)
-        mock_data_from_db.return_value = {"offer_acronym": training.entity_identity.acronym, "offer_type": training.type.name}
-        fake_training_repository = get_fake_training_repository([training])
-        result = CalculateEndPostponement.calculate_program_tree_end_postponement(
-            identity=self.program_tree.entity_id,
-            training_repository=fake_training_repository,
-            mini_training_repository=None
-        )
-        self.assertEqual(result, self.maximum_postponement_year)
-
-    @mock.patch("program_management.ddd.domain.service.identity_search._get_data_from_db")
-    def test_when_root_node_is_training_and_end_year_gt_maximum_postponement(self, mock_data_from_db):
-        end_year = self.maximum_postponement_year + 5
-        training = TrainingFactory(entity_identity__year=self.current_year, end_year=end_year)
-        mock_data_from_db.return_value = {"offer_acronym": training.entity_identity.acronym, "offer_type": training.type.name}
-        fake_training_repository = get_fake_training_repository([training])
-        result = CalculateEndPostponement.calculate_program_tree_end_postponement(
-            identity=self.program_tree.entity_id,
-            training_repository=fake_training_repository,
-            mini_training_repository=None
-        )
-        self.assertEqual(result, self.maximum_postponement_year)
-
-    @mock.patch("program_management.ddd.domain.service.identity_search._get_data_from_db")
-    def test_when_root_node_is_training_and_end_year_lt_maximum_postponement(self, mock_data_from_db):
-        end_year = self.maximum_postponement_year - 3
-        training = TrainingFactory(entity_identity__year=self.current_year, end_year=end_year)
-        mock_data_from_db.return_value = {"offer_acronym": training.entity_identity.acronym, "offer_type": training.type.name}
-        fake_training_repository = get_fake_training_repository([training])
-        result = CalculateEndPostponement.calculate_program_tree_end_postponement(
-            identity=self.program_tree.entity_id,
-            training_repository=fake_training_repository,
-            mini_training_repository=None
-        )
-        self.assertEqual(result, end_year)
-
-    @mock.patch("program_management.ddd.domain.service.identity_search._get_data_from_db")
-    def test_when_root_node_is_training_and_end_year_equals_maximum_postponement(self, mock_data_from_db):
-        end_year = self.maximum_postponement_year
-        training = TrainingFactory(entity_identity__year=self.current_year, end_year=end_year)
-        mock_data_from_db.return_value = {"offer_acronym": training.entity_identity.acronym, "offer_type": training.type.name}
-        fake_training_repository = get_fake_training_repository([training])
-        result = CalculateEndPostponement.calculate_program_tree_end_postponement(
-            identity=self.program_tree.entity_id,
-            training_repository=fake_training_repository,
-            mini_training_repository=None
-        )
-        self.assertEqual(result, self.maximum_postponement_year)
+        self.assertTrue(mock_domain_service.called)
